+++ frontend/playwright/ui/specs/tokens/apply.spec.js
@@ -82,7 +82,7 @@ test.describe("Tokens: Apply token", () => {
     await brTokenPillSM.click();
 
     // Change token from dropdown
-    const brTokenOptionXl = borderRadiusSection.getByLabel("borderRadius.xl");
+    const brTokenOptionXl = borderRadiusSection.getByRole('option', { name: 'borderRadius.xl' })
     await expect(brTokenOptionXl).toBeVisible();
     await brTokenOptionXl.click();
 
@@ -148,7 +148,7 @@ test.describe("Tokens: Apply token", () => {
     await detachButton.click();
 
     // Open dropdown from input
-    const dropdownBtn = layerMenuSection.getByLabel("Open token list");
+    const dropdownBtn = layerMenuSection.getByRole('button', { name: 'Open token list' })
     await expect(dropdownBtn).toBeVisible();
     await dropdownBtn.click();
 
@@ -224,8 +224,8 @@ test.describe("Tokens: Apply token", () => {
       await expect(firstShadowFields).toBeVisible();
 
       // Fill in the shadow values
-      const offsetXInput = firstShadowFields.getByLabel("X");
-      const offsetYInput = firstShadowFields.getByLabel("Y");
+      const offsetXInput = firstShadowFields.getByRole('textbox', { name: 'X' });
+      const offsetYInput = firstShadowFields.getByRole('textbox', { name: 'Y' });
       const blurInput = firstShadowFields.getByRole("textbox", {
         name: "Blur",
       });
@@ -298,8 +298,8 @@ test.describe("Tokens: Apply token", () => {
       await expect(thirdShadowFields).toBeVisible();
 
       // User adds values for the third shadow
-      const thirdOffsetXInput = thirdShadowFields.getByLabel("X");
-      const thirdOffsetYInput = thirdShadowFields.getByLabel("Y");
+      const thirdOffsetXInput = thirdShadowFields.getByRole('textbox', { name: 'X' });
+      const thirdOffsetYInput = thirdShadowFields.getByRole('textbox', { name: 'Y' });
       const thirdBlurInput = thirdShadowFields.getByRole("textbox", {
         name: "Blur",
       });
@@ -327,10 +327,10 @@ test.describe("Tokens: Apply token", () => {
 
       // Verify that the first shadow kept its values
       const firstOffsetXValue = await firstShadowFields
-        .getByLabel("X")
+        .getByRole('textbox', { name: 'X' })
         .inputValue();
       const firstOffsetYValue = await firstShadowFields
-        .getByLabel("Y")
+        .getByRole('textbox', { name: 'Y' })
         .inputValue();
       const firstBlurValue = await firstShadowFields
         .getByRole("textbox", { name: "Blur" })
@@ -356,10 +356,10 @@ test.describe("Tokens: Apply token", () => {
       await expect(newSecondShadowFields).toBeVisible();
 
       const secondOffsetXValue = await newSecondShadowFields
-        .getByLabel("X")
+        .getByRole('textbox', { name: 'X' })
         .inputValue();
       const secondOffsetYValue = await newSecondShadowFields
-        .getByLabel("Y")
+        .getByRole('textbox', { name: 'Y' })
         .inputValue();
       const secondBlurValue = await newSecondShadowFields
         .getByRole("textbox", { name: "Blur" })
@@ -409,10 +409,10 @@ test.describe("Tokens: Apply token", () => {
 
       // Verify first shadow values are still there
       const restoredFirstOffsetX = await firstShadowFields
-        .getByLabel("X")
+        .getByRole('textbox', { name: 'X' })
         .inputValue();
       const restoredFirstOffsetY = await firstShadowFields
-        .getByLabel("Y")
+        .getByRole('textbox', { name: 'Y' })
         .inputValue();
       const restoredFirstBlur = await firstShadowFields
         .getByRole("textbox", { name: "Blur" })
@@ -432,10 +432,10 @@ test.describe("Tokens: Apply token", () => {
 
       // Verify second shadow values are still there
       const restoredSecondOffsetX = await newSecondShadowFields
-        .getByLabel("X")
+        .getByRole('textbox', { name: 'X' })
         .inputValue();
       const restoredSecondOffsetY = await newSecondShadowFields
-        .getByLabel("Y")
+        .getByRole('textbox', { name: 'Y' })
         .inputValue();
       const restoredSecondBlur = await newSecondShadowFields
         .getByRole("textbox", { name: "Blur" })
@@ -517,7 +517,7 @@ test.describe("Tokens: Apply token", () => {
     await dimensionSMTokenPill.nth(1).click();
 
     // Change token from dropdown
-    const dimensionTokenOptionXl = measuresSection.getByLabel("dimension.xl");
+    const dimensionTokenOptionXl = measuresSection.getByRole('option', { name: 'dimension.xl' })
     await expect(dimensionTokenOptionXl).toBeVisible();
     await dimensionTokenOptionXl.click();
 
@@ -571,7 +571,7 @@ test.describe("Tokens: Apply token", () => {
     await dimensionSMTokenPill.click();
 
     // Change token from dropdown
-    const dimensionTokenOptionXl = measuresSection.getByLabel("dimension.xl");
+    const dimensionTokenOptionXl = measuresSection.getByRole('option', { name: 'dimension.xl' });
     await expect(dimensionTokenOptionXl).toBeVisible();
     await dimensionTokenOptionXl.click();
 
@@ -625,7 +625,7 @@ test.describe("Tokens: Apply token", () => {
     await dimensionSMTokenPill.click();
 
     // Change token from dropdown
-    const dimensionTokenOptionXl = measuresSection.getByLabel("dimension.xl");
+    const dimensionTokenOptionXl = measuresSection.getByRole('option', { name: 'dimension.xl' });
     await expect(dimensionTokenOptionXl).toBeVisible();
     await dimensionTokenOptionXl.click();
 
@@ -681,7 +681,7 @@ test.describe("Tokens: Apply token", () => {
 
     // Change token from dropdown
     const dimensionTokenOptionXl =
-      borderRadiusSection.getByLabel("dimension.xl");
+      borderRadiusSection.getByRole('option', { name: 'dimension.xl' });
     await expect(dimensionTokenOptionXl).toBeVisible();
     await dimensionTokenOptionXl.click();
 
@@ -750,7 +750,7 @@ test.describe("Tokens: Apply token", () => {
     });
     await tokenDropdown.click();
 
-    const widthOptionSmall = firstStrokeRow.getByLabel("width-small");
+    const widthOptionSmall = firstStrokeRow.getByRole('option', { name: 'width-small' });
     await expect(widthOptionSmall).toBeVisible();
     await widthOptionSmall.click();
     const StrokeWidthPillSmall = firstStrokeRow.getByRole("button", {
@@ -830,15 +830,10 @@ test.describe("Tokens: Apply token", () => {
     });
     await detachButton.click();
     await expect(marginPillXL).not.toBeVisible();
-    const horizontalMarginInput = layoutItemSectionSidebar.getByText(
-      "Horizontal marginOpen token",
-    );
-    await expect(horizontalMarginInput).toBeVisible();
-
-    const tokenDropdown = horizontalMarginInput.getByRole("button", {
+    const horizontalMarginInput = layoutItemSectionSidebar.getByRole("button", {
       name: "Open token list",
     });
-    await tokenDropdown.click();
+    await horizontalMarginInput.nth(1).click();
 
     await expect(dimensionTokenOptionXl).toBeVisible();
     await dimensionTokenOptionXl.click();