+++ frontend/src/app/main/ui/settings/subscription.cljs
@@ -389,13 +389,10 @@
 
 (mf/defc subscription-page*
   [{:keys [profile]}]
-  (let [route          (mf/deref refs/route)
-        authenticated? (da/is-authenticated? profile)
-
-        nitrate-license (dm/get-in profile [:props :nitrate-license])
-
-        nitrate? (and (contains? cf/flags :nitrate)
-                      (:valid nitrate-license))
+  (let [route           (mf/deref refs/route)
+        authenticated?  (da/is-authenticated? profile)
+        nitrate-license (:subscription profile)
+        nitrate?        (dnt/is-valid-license? profile)
 
         params-subscription
         (-> route :params :query :subscription)