+++ frontend/src/app/main/ui/workspace/sidebar/options/rows/color_row.cljs
@@ -94,16 +94,17 @@
         not-active (or (empty? active-tokens)
                        (nil? token))
         id (dm/str (:id token) "-name")
+        token-name-ref (mf/use-ref nil)
         swatch-tooltip-content (cond
                                  not-active
-                                 (tr "ds.inputs.token-field.no-active-token-option" token-name)
+                                 (tr "ds.inputs.token-field.no-active-color.token-option")
                                  has-errors
                                  (tr "color-row.token-color-row.deleted-token")
                                  :else
                                  (tr "workspace.tokens.resolved-value" resolved))
         name-tooltip-content (cond
                                not-active
-                               (tr "ds.inputs.token-field.no-active-token-option" token-name)
+                               (tr "ds.inputs.token-field.no-active-color.token-option")
                                has-errors
                                (tr "color-row.token-color-row.deleted-token")
                                :else
@@ -126,8 +127,11 @@
                     :size "small"}]]
       [:> tooltip* {:content name-tooltip-content
                     :id id
+                    :aria-label (str (tr "workspace.tokens.token-name") ": " applied-token-name)
+                    :trigger-ref token-name-ref
                     :class (stl/css :token-tooltip)}
        [:div {:class (stl/css :token-name)
+              :ref token-name-ref
               :aria-labelledby id}
         (or token-name applied-token-name)]]
       [:div {:class (stl/css :token-actions)}